pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "RAG Pipeline with Anthropic Claude & Voyage AI"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "jaxcoder", email = "jromero@jaxlabs.xyz"}
]

dependencies = [
    "python-dotenv>=1.0.0",
    "anthropic>=0.18.1",
    "voyageai>=0.2.1",
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    "scikit-learn>=1.3.2",
    "chromadb>=0.4.22",
    "tiktoken>=0.5.2",
    "pypdf>=3.17.4",
    "beautifulsoup4>=4.12.2",
    "markdown>=3.5.1",
    "aiohttp>=3.9.1",
    "requests>=2.31.0",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "ipython>=8.17.2",
    "ipdb>=0.13.13",
    "jupyter>=1.0.0",
]

[project.scripts]
rag-cli = "src.cli:main"
rag-example = "src.example_usage:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q --strict-markers --cov=src"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "ipython>=8.17.2",
    "ipdb>=0.13.13",
    "jupyter>=1.0.0",
]
